backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "backend for my project"
authors = ["praveen"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
pillow = "^11.0.0"
torch = "^2.5.1"
transformers = "^4.47.1"
pymilvus = "^2.5.2"
jinja2 = "^3.1.5"
sentencepiece = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
